heap_allocator: Ensure that enough space is allocted for the free list
Found a bug were the heap allocator page faults when the allocation size
is less than the size of a free list entry. Then the free list will be
accidently modified when adjusting the tags during spliting, leading to
failures when trying to replace it with the new free object.

As a part of this commit, minor formating changes were made to comply
with the mandated comment style.