fs: Allow open to both set an inode and return status
When previously changing open to return an inode pointer it made calls
to open unable to disguish between if the file didn't exits or an io
error had occured. This commit address this by re-introducing an int
return value and adds a pointer to the inode pointer in the argument
list.