inode: Revert get_inode api back to returning an inode pointer
The original api change was made to reduce the amount of error handling
code need for fs implementation when doing path traversal. However,
since we're about to introduce a path walker that gives the traversal
responsibilty to the vfs that consideration is no longert need and
having the old api would simplify the path walker code.