fs: Replace vfs nodes with nested mounting through superblocks
The vfs node system made it hard to implement a root fs, so a major
re-write was made thorugh out the vfs nodes a replace the with
superblocks that could be mounted to exisiting inodes.

In order to allow this change a pathwalker was created, so that the vfs
and not the fs implementation handled the path resolution. Once it's
handled by the vfs, it's easy to handle mountpoints by simply replacing
the inodes once necesary.

Should proably be split it smaller commits, but well it's a hobby
project :)