utils: Create a GET_STRUCT macro
The GET_STRUCT macro allows one to compute the pointer to a struct based
on a pointer to one of its member. Useful when represting inheritence
with nested struct, with this macro one can easily access the the child
struct from the parent embedded into the child as a field, e.g.

struct child {
    ...
    struct parent p;
};

child_ptr = GET_STRUCT(struct child, p, parent_ptr);