fs: Replace sysfs with pseudo_file based kinfofs
Replace sysfs with kinfos which uses pseudo_file objects instead of
hard-coded lists to represent the files and directories. This allows for
greater flexiblity and makes it a more powerful fs in general.

The name change from sysfs to kinfo, was made to clarify it's purpose,
namely to supply the userspace with kernel information.