paging.h: Typecast addresses to uintptr_t within the L2P/P2L macros
If one supplies pointer to a struct to the L2P/P2L macros it will treat
the addition and subtractions as multiples of struct size, leading to
unexpected address.

To address this problem, the addresses provided to the macros is now
typecasted into uintptr_t before applying the arithmetic.