i386: kernel_thread_switch: Apply TSS offset to field and not ptr
Adding the offset to the pointer, results in the movl storting the value
of ebx into address after the pointer, which lead to it accidentally
overwritting other parts of the global data.

To fix this one could do either use the indexed  mov notation or
dereference it first, but that would involve a register. The simply
solution is to just to remove the pointer and use the global data symbol
directly.