devices: Initial device and driver framework
Creates the initial device and driver framework by defining a objects
for drivers and device instances. To keep things simple, we have single
static table with all drivers and an api to dynamically register devices
to the drivers.

This patch also adds some devfs glue, by allowing a device to be read
based on its device number.