tty: Move key state from kbd driver to tty
Since the tty need to store some kind kbd state in order to be able to
restore it with tty switches, we save some duplicaiton by moving it
there.

This commit also re-define how keycode are specified in order reduce the
number of operations needed to set and read them.