list: Make circular with a sentinel node
Re-write list to be a circular linked list with the list struct acting
as sentinel/dummy node. This approach makes list manipulation faster
since we avoid the need of doing NULL checking.

The commit also introduces functions to add/remove nodes both at the
head and the tail of the list, to allow it to replace exisitng queue
implementations.