tasks/locking: Only disable interrupts for mutexes before scheduler init
Previously, interrupts was enabled/disabled before scheduler
initialisation for semaphores, and indirectly for mutexes; this is
incorrect, since semaphores isn't necesarly used as a acquire/release
pair. So, move it where it belongs, into the mutex.

Mutex should be used pre-scheduler initialisation, so that is to be
fixed in a future commit.